match marioAccess

+++ src/Game/Player/MarioActorMorph.cpp
@@ -15,7 +15,7 @@
 #include "Game/Util/SoundUtil.hpp"
 #include "revolution/types.h"
 
-void MarioActor::setPlayerMode(u8 playerMode, bool myBool) {
+void MarioActor::setPlayerMode(u16 playerMode, bool myBool) {
     if (mPlayerMode == (u16)playerMode) {
         if (playerMode == 1) {
             _3DC = mConst->getTable()->mMetalHoldTime;
